+++ src/graph/backend/dnnl/passes/transform.cpp
@@ -2790,8 +2790,9 @@ status_t fuse_adjacent_reorders(std::shared_ptr<subgraph_t> &sg) {
             fused_dst_zps.reserve(max_num);
             for (size_t i = 0; i < max_num; i++) {
                 fused_scales.emplace_back(scales1[i] * scales2[i]);
-                fused_dst_zps.emplace_back(
-                        scales2[i] * (dst_zps1[i] - src_zps2[i]) + dst_zps2[i]);
+                fused_dst_zps.emplace_back(static_cast<int64_t>(
+                        scales2[i] * (dst_zps1[i] - src_zps2[i])
+                        + dst_zps2[i]));
             }
 
             int64_t axis = -1;
